Adding create user endpoint for admin, refactor auth signUn
