+++ server/package.json
@@ -40,7 +40,6 @@
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
     "nanoid": "^5.0.9",
-    "pg": "^8.13.3",
     "winston": "^3.17.0"
   }
 }
