--- server/src/db.ts
@@ -1,11 +1,0 @@
-import { Pool } from "pg";
-
-const dbPool = new Pool({
-  user: process.env.DB_USER,
-  host: process.env.DB_HOST,
-  database: process.env.DB_NAME,
-  password: process.env.DB_PASSWORD,
-  port: process.env.DB_PORT ? parseInt(process.env.DB_PORT): undefined,
-});
-
-export default dbPool;
