+++ server/src/utils/format-user.ts
@@ -1,4 +1,3 @@
-// src/utils/formatUser.ts
 import { IUser } from '../models/interfaces';
 import { users as PrismaUser } from '../generated/prisma';
 
@@ -9,8 +8,6 @@ export function formatUser(
     id: user.id,
     nickname: user.nickname,
     email: user.email,
-    hashed_code: user.hashed_code,
-    role_id: user.role_id,
     role: {
       id: user.user_roles.id,
       name: user.user_roles.name,
